SORQA-982 : v2 test with tab handle
