+++ sormas-e2e-tests/src/test/java/org/sormas/e2etests/steps/web/application/aCommonComponents/GeneralActionsSteps.java
@@ -56,5 +56,6 @@ public class GeneralActionsSteps implements En {
               " text is not present in handover component");
           softly.assertAll();
         });
+    When("I refresh current page", () -> webDriverHelpers.refreshCurrentPage());
   }
 }
