+++ sormas-e2e-tests/src/test/java/org/sormas/e2etests/steps/web/application/shares/SharesDirectorySteps.java
@@ -18,9 +18,10 @@
 
 package org.sormas.e2etests.steps.web.application.shares;
 
-import static org.sormas.e2etests.pages.application.cases.CaseDirectoryPage.ACTION_OKAY;
+import static org.sormas.e2etests.pages.application.cases.CaseDirectoryPage.*;
 import static org.sormas.e2etests.pages.application.shares.EditSharesPage.*;
 
+import com.github.javafaker.Faker;
 import cucumber.api.java8.En;
 import java.util.concurrent.TimeUnit;
 import javax.inject.Inject;
@@ -33,18 +34,77 @@ import org.testng.asserts.SoftAssert;
 public class SharesDirectorySteps implements En {
 
   private final WebDriverHelpers webDriverHelpers;
+  public static Faker faker;
+  private static String generatedRandomStringCase;
+  private static String generatedRandomStringContact;
 
   @Inject
   public SharesDirectorySteps(
-      WebDriverHelpers webDriverHelpers, SoftAssert softly, ApiState apiState) {
+      WebDriverHelpers webDriverHelpers, SoftAssert softly, ApiState apiState, Faker faker) {
     this.webDriverHelpers = webDriverHelpers;
+    this.faker = faker;
 
     When(
         "I accept first entity from table in Shares Page",
         () -> {
-          webDriverHelpers.waitUntilIdentifiedElementIsVisibleAndClickable(ACCEPT_BUTTON);
           webDriverHelpers.clickOnWebElementBySelector(ACCEPT_BUTTON);
           TimeUnit.SECONDS.sleep(2); // wait for results to reload
+        });
+
+    When(
+        "I click on {string} shared contact button with copied contact description",
+        (String option) -> {
+          switch (option) {
+            case "reject":
+              webDriverHelpers.clickOnWebElementBySelector(
+                  getActionRejectButtonByContactDescription(generatedRandomStringContact));
+              break;
+            case "accept":
+              webDriverHelpers.clickOnWebElementBySelector(
+                  getActionAcceptButtonByContactDescription(generatedRandomStringContact));
+              break;
+          }
+        });
+
+    When(
+        "I click on {string} shared case button with copied case description",
+        (String option) -> {
+          switch (option) {
+            case "reject":
+              webDriverHelpers.clickOnWebElementBySelector(
+                  getActionRejectButtonByCaseDescription(generatedRandomStringCase));
+              break;
+            case "accept":
+              webDriverHelpers.clickOnWebElementBySelector(
+                  getActionAcceptButtonByCaseDescription(generatedRandomStringCase));
+              break;
+          }
+        });
+
+    And(
+        "^I check that accept shared case button with copied case description is visible in Share Directory page$",
+        () -> {
+          webDriverHelpers.waitUntilIdentifiedElementIsPresent(
+              getActionAcceptButtonByCaseDescription(generatedRandomStringCase));
+        });
+
+    And(
+        "^I check that accept shared contact button with copied contact description is visible in Share Directory page$",
+        () -> {
+          webDriverHelpers.waitUntilIdentifiedElementIsPresent(
+              getActionAcceptButtonByContactDescription(generatedRandomStringContact));
+        });
+
+    And(
+        "^I check that entity not found error popup is displayed in Share Directory page$",
+        () -> {
+          webDriverHelpers.waitUntilIdentifiedElementIsPresent(ENTITY_NOT_FOUND_POPUP);
+        });
+
+    And(
+        "I fill comment in share popup with {string}",
+        (String comment) -> {
+          webDriverHelpers.fillInWebElement(EXTRA_COMMENT_INPUT_SHARE_POPUP, comment);
         });
 
     When(
@@ -54,6 +114,22 @@ public class SharesDirectorySteps implements En {
               webDriverHelpers.isElementVisibleWithTimeout(ACCEPT_BUTTON, 3),
               "Accept button is visible!");
           softly.assertAll();
+        });
+
+    And(
+        "I fill comment in share popup for case with random string",
+        () -> {
+          generatedRandomStringCase = faker.beer().name();
+          webDriverHelpers.fillInWebElement(
+              EXTRA_COMMENT_INPUT_SHARE_POPUP, generatedRandomStringCase);
+        });
+
+    And(
+        "I fill comment in share popup for contact with random string",
+        () -> {
+          generatedRandomStringContact = faker.beer().name();
+          webDriverHelpers.fillInWebElement(
+              EXTRA_COMMENT_INPUT_SHARE_POPUP, generatedRandomStringContact);
         });
 
     When(
